+++ src/pages/dashboard.js
@@ -10,11 +10,9 @@ import Export from './util/export';
  */
 async function initTable(table) {
   try {
-    const storage = await browser.storage.sync.get();
-
-    const bookmarkList = Object.keys(storage)
-      .filter(key => (storage[key].type && storage[key].type === 'bookmark'))
-      .map(key => storage[key]);
+    const bookmarkList = await browser.runtime.sendMessage({
+      type: 'get-bookmark-data',
+    });
 
     const promises = [];
     bookmarkList.forEach(async (bookmark) => {