+++ src/util/upgrade.js
@@ -1,5 +1,5 @@
 import SourceMigrations from './source-migration';
-import store from '../util/datastore';
+import * as FoxyStorage from '../util/FoxyStorage';
 
 /**
  * Checks whether or not v1 is lower than v2.
@@ -60,13 +60,13 @@ export function applyChanges(object, changes) {
  */
 async function updateStorage(bookmark, changes) {
   try {
-    const manga = await store.getItem(`${bookmark.source}/${bookmark.reference}`);
+    const manga = await FoxyStorage.DataStorage.getItem(`${bookmark.source}/${bookmark.reference}`);
     if (!manga) Promise.resolve(false);
 
     const newMangaObj = applyChanges(manga, changes);
 
-    await store.removeItem(`${bookmark.source}/${bookmark.reference}`);
-    await store.setItem(`${newMangaObj.source}/${newMangaObj.reference}`, newMangaObj);
+    await FoxyStorage.DataStorage.removeItem(`${bookmark.source}/${bookmark.reference}`);
+    await FoxyStorage.DataStorage.setItem(`${newMangaObj.source}/${newMangaObj.reference}`, newMangaObj);
 
     return Promise.resolve(true);
   } catch (err) {
@@ -100,14 +100,13 @@ async function migrateDB(previousVersion, bookmarkList, updateDb = true) {
       const bookmark = bookmarkList[i];
       const key = `${bookmark.source}/${bookmark.reference}`;
 
-      const storage = {};
-      storage[key] = applyChanges(bookmark, changes);
+      const updatedEntry = applyChanges(bookmark, changes);
 
-      newList.push(storage[key]);
+      newList.push(updatedEntry);
 
       // Update indexedDB storage
       if (updateDb) {
-        promises.push(browser.storage.sync.set(storage));
+        promises.push(FoxyStorage.setMetadata(key, updatedEntry));
         promises.push(updateStorage(bookmark, changes));
       }
     }
@@ -135,9 +134,7 @@ async function redesignBookmarkStorage(bookmarkList) {
 
       const newBookmark = Object.assign({}, bookmark, { type: 'bookmark' });
 
-      const storage = {};
-      storage[key] = newBookmark;
-      promises.push(browser.storage.sync.set(storage));
+      promises.push(FoxyStorage.setMetadata(key, newBookmark));
     });
 
     await Promise.all(promises);
@@ -156,7 +153,7 @@ async function redesignBookmarkStorage(bookmarkList) {
  */
 export async function updateAddon(previousVersion) {
   try {
-    let storage = await browser.storage.sync.get();
+    const storage = await browser.storage.sync.get();
 
     // Handle transition to v0.3.1 - inclusion of manga list view mode
     if (compareVersions(previousVersion, '0.3.1') && typeof storage.view_mode === 'string') {
@@ -168,12 +165,15 @@ export async function updateAddon(previousVersion) {
     // Handle transition to v0.6.0 - storage redesign (issue#9)
     if (compareVersions(previousVersion, '0.6.0') && storage.bookmark_list) {
       await redesignBookmarkStorage(storage.bookmark_list);
-      storage = await browser.storage.sync.get();
     }
 
-    const bookmarkList = Object.keys(storage)
-      .filter(key => (storage[key].type && storage[key].type === 'bookmark'))
-      .map(key => storage[key]);
+    const bookmarkList = await FoxyStorage.getMetadata();
+
+    // Handle transition to v0.6.2 - local/sync storage options (issue#13)
+    if (compareVersions(previousVersion, '0.6.2') && bookmarkList.length > FoxyStorage.syncLimit) {
+      await FoxyStorage.switchStorage('local');
+      await browser.storage.sync.set({ storageType: 'local' });
+    }
 
     // Applies all DB migrations needed
     await migrateDB(previousVersion, bookmarkList);